Relocated Method and deleted unecesary class file
